Added build info health check
Added build info Dropwizard health check to display the build number as
defined by the implementation version in the manifest in the fat JAR.

The build info health check takes a Java Package object as argument.

The build number will be set to "[not available]" when the
implementation version is not available, for example when the service is
being run locally from the source folder using 'sbt run'.

Includes unit tests and documentation. Switched to given-when-then unit test
style for this modification.